requirements.txt
Flask==2.2.2
Pillow==9.4.0
opencv-python>=4.7.0
numpy>=1.24.0
